+++ packages/paste-core/components/card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/card@1.3.4...@twilio-paste/card@1.3.5) (2020-02-11)
+
+**Note:** Version bump only for package @twilio-paste/card
+
+
+
+
+
+## [1.3.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/card@1.3.3...@twilio-paste/card@1.3.4) (2020-02-07)
+
+**Note:** Version bump only for package @twilio-paste/card
+
+
+
+
+
 ## [1.3.3](https://github.com/twilio-labs/paste/compare/@twilio-paste/card@1.3.2...@twilio-paste/card@1.3.3) (2020-02-03)
 
 **Note:** Version bump only for package @twilio-paste/card