+++ packages/paste-core/utilities/grid/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.6](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.1.5...@twilio-paste/grid@0.1.6) (2020-02-11)
+
+**Note:** Version bump only for package @twilio-paste/grid
+
+
+
+
+
+## [0.1.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.1.4...@twilio-paste/grid@0.1.5) (2020-02-10)
+
+
+### Bug Fixes
+
+* **grid:** add getColumnGutters function to process token value ([#298](https://github.com/twilio-labs/paste/issues/298)) ([684b7dc](https://github.com/twilio-labs/paste/commit/684b7dc18a5a12b3f5862bd201518912113a1a72))
+
+
+
+
+
 ## [0.1.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/grid@0.1.3...@twilio-paste/grid@0.1.4) (2020-02-05)
 
 **Note:** Version bump only for package @twilio-paste/grid